projects/oblique/src/lib/service-navigation/service-navigation.component.ts
import {Component, Input, OnInit, ViewEncapsulation} from '@angular/core';
import {ObServiceNavigationService} from './service-navigation.service';
import {ObEPamsEnvironment} from './service-navigation.model';

@Component({
	selector: 'ob-service-navigation',
	templateUrl: './service-navigation.component.html',
	styleUrls: ['./service-navigation.component.scss'],
	providers: [ObServiceNavigationService],
	encapsulation: ViewEncapsulation.None,
	host: {class: 'ob-service-navigation'}
})
export class ObServiceNavigationComponent implements OnInit {
	@Input() environment: ObEPamsEnvironment;
	@Input() rootUrl: string;
	readonly loginUrl$ = this.headerControlsService.getLoginUrl$();

	constructor(private readonly headerControlsService: ObServiceNavigationService) {}

	ngOnInit(): void {
		this.headerControlsService.setUpRootUrls(this.environment, this.rootUrl);
	}
}
